Moved flame to chifra
