+++ src/apps/chifra/cmd/serve/options.go
@@ -1,4 +1,4 @@
-package server
+package serve
 
 /*-------------------------------------------------------------------------------------------
  * qblocks - fast, easily-accessible, fully-decentralized data from blockchains
@@ -14,15 +14,8 @@ package server
  *-------------------------------------------------------------------------------------------*/
 
 import (
-	"encoding/json"
 	"fmt"
 	"log"
-	"os"
-	"os/exec"
-	"strings"
-
-	utils "github.com/TrueBlocks/trueblocks-core/src/go-apps/blaze/utils"
-	flag "github.com/spf13/pflag"
 )
 
 // ChifraStatus
@@ -58,22 +51,24 @@ type ChifraResponse struct {
 }
 
 func GetChifraResponse() (ChifraResponse, error) {
-	cmd := exec.Command(GetCommandPath("cacheStatus"), "--terse")
-	cmd.Env = append(os.Environ(), "API_MODE=true")
-	out, err := cmd.Output()
-	if err != nil {
-		fmt.Printf("ChifraResponse:cmd.Output() failed: %s", err)
-		var junk ChifraResponse
-		return junk, err
-	}
-	var response ChifraResponse
-	err = json.Unmarshal(out, &response)
-	if err != nil {
-		fmt.Printf("ChifraResponse:json.Unmarshal() failed: %s", err)
-		var junk ChifraResponse
-		return junk, err
-	}
-	return response, nil
+	// TODO: fix
+	// cmd := exec.Command(GetCommandPath("cacheStatus"), "--terse")
+	// cmd.Env = append(os.Environ(), "API_MODE=true")
+	// out, err := cmd.Output()
+	// if err != nil {
+	// 	fmt.Printf("ChifraResponse:cmd.Output() failed: %s", err)
+	// 	var junk ChifraResponse
+	// 	return junk, err
+	// }
+	// var response ChifraResponse
+	// err = json.Unmarshal(out, &response)
+	// if err != nil {
+	// 	fmt.Printf("ChifraResponse:json.Unmarshal() failed: %s", err)
+	// 	var junk ChifraResponse
+	// 	return junk, err
+	// }
+	// return response, nil
+	return ChifraResponse{}, nil
 }
 
 func GetChifraData() (ChifraStatus, error) {
@@ -117,38 +112,38 @@ var Options OptionsType
 
 func ParseOptions() error {
 	// Establish and parse the command line input...
-	flag.StringVar(&Options.Port, "port", ":8080", "specify the server's port")
-	if !strings.HasPrefix(Options.Port, ":") {
-		Options.Port = ":" + Options.Port
-	}
+	// flag.StringVar(&Options.Port, "port", ":8080", "specify the server's port")
+	// if !strings.HasPrefix(Options.Port, ":") {
+	// 	Options.Port = ":" + Options.Port
+	// }
 
-	flag.IntVar(&Options.Verbose, "verbose", 0, "verbose level (between 0 and 10 inclusive)")
-	flag.BoolVar(&Options.Pin, "pin", false, "pins Bloom filters and chunks to pinning service (requires API key)")
-	flag.IntVar(&Options.Sleep, "sleep", 14, "specifies sleep interval between scrapes")
+	// flag.IntVar(&Options.Verbose, "verbose", 0, "verbose level (between 0 and 10 inclusive)")
+	// flag.BoolVar(&Options.Pin, "pin", false, "pins Bloom filters and chunks to pinning service (requires API key)")
+	// flag.IntVar(&Options.Sleep, "sleep", 14, "specifies sleep interval between scrapes")
 
-	flag.Parse()
+	// flag.Parse()
 
 	Options.Status, _ = GetChifraData()
 	Options.Meta, _ = GetChifraMeta()
 
 	if Options.Port != ":8080" {
-		log.Print(utils.Green, "port:        ", utils.Off, Options.Port, "\n")
+		log.Print( /*utils.Green, */ "port:        " /*utils.Off,*/, Options.Port, "\n")
 	} else {
 		Options.Port = "127.0.0.1:8080"
 	}
 
-	if Options.Verbose > 0 {
-		log.Print(utils.Green, "verbose:     ", utils.Off, Options.Verbose, "\n")
-	}
+	// if Options.Verbose > 0 {
+	// 	log.Print(utils.Green, "verbose:     ", utils.Off, Options.Verbose, "\n")
+	// }
 
-	log.Print("\n")
-	log.Print(utils.Green, "Starting API server on port "+Options.Port, utils.Off, "\n")
-	log.Print(utils.Green, "Client:       ", utils.Off, Options.Status.Client)
-	log.Print(utils.Green, "TrueBlocks:   ", utils.Off, Options.Status.TrueBlocks)
-	log.Print(utils.Green, "Cache Path:   ", utils.Off, Options.Status.CachePath)
-	log.Print(utils.Green, "Index Path:   ", utils.Off, Options.Status.IndexPath)
-	log.Print(utils.Green, "Rpc Provider: ", utils.Off, Options.Status.RPC)
-	log.Print(utils.Green, "Progress:     ", utils.Off, Options.Meta.Client, ", ", Options.Meta.Finalized, ", ", Options.Meta.Staging, ", ", Options.Meta.Unripe)
+	// log.Print("\n")
+	// log.Print(utils.Green, "Starting API server on port "+Options.Port, utils.Off, "\n")
+	// log.Print(utils.Green, "Client:       ", utils.Off, Options.Status.Client)
+	// log.Print(utils.Green, "TrueBlocks:   ", utils.Off, Options.Status.TrueBlocks)
+	// log.Print(utils.Green, "Cache Path:   ", utils.Off, Options.Status.CachePath)
+	// log.Print(utils.Green, "Index Path:   ", utils.Off, Options.Status.IndexPath)
+	// log.Print(utils.Green, "Rpc Provider: ", utils.Off, Options.Status.RPC)
+	// log.Print(utils.Green, "Progress:     ", utils.Off, Options.Meta.Client, ", ", Options.Meta.Finalized, ", ", Options.Meta.Staging, ", ", Options.Meta.Unripe)
 
 	return nil
 }
