+++ src/apps/chifra/cmd/serve_run.go
@@ -13,8 +13,11 @@
 package cmd
 
 import (
+	"log"
+	"net/http"
 	"strings"
 
+	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/cmd/serve"
 	"github.com/TrueBlocks/trueblocks-core/src/apps/chifra/pkg/validate"
 	"github.com/spf13/cobra"
 )
@@ -28,13 +31,26 @@ func validateServeArgs(cmd *cobra.Command, args []string) error {
 }
 
 func runServe(cmd *cobra.Command, args []string) {
-	options := ":8080"
-	if len(ServeOpts.port) > 0 {
-		options = " --port " + ServeOpts.port
-	}
+	// options := ":8080"
+	// if len(ServeOpts.port) > 0 {
+	// 	options = " --port " + ServeOpts.port
+	// }
 	arguments := ""
 	for _, arg := range args {
 		arguments += " " + arg
 	}
-	PassItOn("flame", options, arguments)
+	// PassItOn("flame", options, arguments)
+
+	// Handle command line options
+	err := serve.ParseOptions()
+	if err != nil {
+		log.Println("Could not parse command line.")
+		return
+	}
+
+	// Start listening on web sockets
+	serve.RunWebsocketPool()
+
+	// Start listening for requests
+	log.Fatal(http.ListenAndServe(serve.Options.Port, serve.NewRouter()))
 }
