+++ src/apps/chifra/go.mod
@@ -6,9 +6,12 @@ require (
 	github.com/BurntSushi/toml v0.4.1
 	github.com/alecthomas/participle/v2 v2.0.0-alpha7
 	github.com/ethereum/go-ethereum v1.10.11
+	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/gorilla/websocket v1.4.2 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/panjf2000/ants/v2 v2.4.6
 	github.com/spf13/cobra v1.2.1
 	github.com/spf13/viper v1.9.0
 	golang.org/x/crypto v0.0.0-20210817164053-32db794688a5
+	golang.org/x/time v0.0.0-20210220033141-f8bda1e9f3ba // indirect
 )
