--- src/go-apps/flame/README.md
@@ -1,43 +1,0 @@
-## chifra serve
-
-`chifra serve` delivers a JSON API for each of the `chifra` commands along with each of its options.
-It does this through `flame` server, which is written in Go.
-
-Another way to get help to run `chifra --help` or `chifra <cmd> --help` on your command line. See below for an example of converting command line options to a call to the API. There's a one-to-one correspondence between the command line tools and options and the API routes and their options.
-
-```[plaintext]
-Purpose:
-  Serve the TrueBlocks API using the flame server.
-
-Usage:
-  chifra serve [flags]
-
-Flags:
-  -p, --port string   specify the server's port (:8080 default)
-
-Global Flags:
-  -x, --fmt string   export format, one of [none|json*|txt|csv|api]
-  -h, --help         display this help screen
-  -v, --verbose      enable verbose (increase detail with --log_level)
-
-Notes:
-  - To start API open terminal window and run chifra serve.
-  - See the API documentation for more information.
-```
-
-Other Options
-
-All tools accept the following additional flags, although in some cases, they have no meaning.
-
-```[plaintext]
-  -v, --version         display the current version of the tool
-      --wei             export values in wei (the default)
-      --ether           export values in ether
-      --dollars         export values in US dollars
-      --raw             pass raw RPC data directly from the node with no processing
-      --to_file         write the results to a temporary file and return the filename
-      --output string   write the results to file 'fn' and return the filename
-      --file string     specify multiple sets of command line options in a file
-```
-
-*For the `--file string` option, you may place a series of valid command lines in a file using any valid flags. In some cases, this may significantly improve performance. A semi-colon at the start of any line makes it a comment.*
