--- src/go-apps/flame/main.go
@@ -1,36 +1,0 @@
-package main
-
-/*-------------------------------------------------------------------------------------------
- * qblocks - fast, easily-accessible, fully-decentralized data from blockchains
- * copyright (c) 2016, 2021 TrueBlocks, LLC (http://trueblocks.io)
- *
- * This program is free software: you may redistribute it and/or modify it under the terms
- * of the GNU General Public License as published by the Free Software Foundation, either
- * version 3 of the License, or (at your option) any later version. This program is
- * distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even
- * the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details. You should have received a copy of the GNU General
- * Public License along with this program. If not, see http://www.gnu.org/licenses/.
- *-------------------------------------------------------------------------------------------*/
-
-import (
-	"log"
-	"net/http"
-
-	server "github.com/TrueBlocks/trueblocks-core/src/go-apps/blaze/server"
-)
-
-func main() {
-	// Handle command line options
-	err := server.ParseOptions()
-	if err != nil {
-		log.Println("Could not parse command line.")
-		return
-	}
-
-	// Start listening on web sockets
-	server.RunWebsocketPool()
-
-	// Start listening for requests
-	log.Fatal(http.ListenAndServe(server.Options.Port, server.NewRouter()))
-}
